update InventoryItem Integration
refactor: File structure for inventory api testing
