+++ backend/src/__tests__/fixtures/inventory/getInventoryItemsFixtures.ts
@@ -1,0 +1,24 @@
+export const testFetchInventoryItems = [
+    {
+        name: "Fishing bait",
+        note: "For tilapias",
+        quantity: 5,
+        unitPrice: 999.99,
+        selectUnit: "pieces",
+        unitSize: 1,
+        total: 4999.95,
+        dateCreated: new Date("2025-01-01"),
+        lastUpdated: new Date("2025-01-02"),
+    },
+    {
+        name: "Fishing Net",
+        note: "24-inch",
+        quantity: 10,
+        unitPrice: 299.99,
+        selectUnit: "pieces",
+        unitSize: 1,
+        total: 2999.90,
+        dateCreated: new Date("2023-01-01"),
+        lastUpdated: null,
+    },
+];
